+++ internal/repo/common/db_fighters_test.go
@@ -1,0 +1,15 @@
+package repo
+
+import "testing"
+
+func TestSearchFightersCount(t *testing.T) {
+	// TODO
+}
+
+func TestSearchFighters(t *testing.T) {
+	// TODO
+}
+
+func TestPerformFightersQuery(t *testing.T) {
+	// TODO
+}