+++ internal/repo/common/pgx_bets_test.go
@@ -1,0 +1,15 @@
+package repo
+
+import "testing"
+
+func TestSearchBetsCount(t *testing.T) {
+	// TODO
+}
+
+func TestSearchBets(t *testing.T) {
+	// TODO
+}
+
+func TestCreateBet(t *testing.T) {
+	// TODO
+}