+++ internal/repo/common/pgx_fights_test.go
@@ -1,0 +1,11 @@
+package repo
+
+import "testing"
+
+func TestTxCreateEventFight(t *testing.T) {
+	// TODO
+}
+
+func TestSetFightResult(t *testing.T) {
+	// TODO
+}